refactor(cypress): consolidate Node-side code under node/
Move cypress/db/ and cypress/shared/ under cypress/node/ so all
Node-process code (config, db access, input mappers, tasks) lives under
one roof, matching the browser/Node split the README describes. Rename
the ambiguous shared/ -> mappers/ so "shared/" now refers only to the
browser-side constants in e2e/shared/.

- cypress/db/         -> cypress/node/db/
- cypress/shared/     -> cypress/node/mappers/
- rewrite 12 import sites in cypress/node/tasks/
- update README directory tree + "Node side" prose

Verified: pnpm typecheck:cypress + pnpm biome:lint clean.
