+++ cypress/node/mappers/user-input.mapper.ts
