+++ cypress/node/tasks/user-exists.task.ts
@@ -1,6 +1,6 @@
-import { nodeDb } from "@cypress/db/node-db";
-import { firstRow } from "@cypress/db/pg-result.utils";
-import { normalizeUserEmail } from "@cypress/shared/user-input.mapper";
+import { nodeDb } from "@cypress/node/db/node-db";
+import { firstRow } from "@cypress/node/db/pg-result.utils";
+import { normalizeUserEmail } from "@cypress/node/mappers/user-input.mapper";
 import { users } from "@database/schema/users";
 import { sql } from "drizzle-orm";
